Add error handling and some workarounds for antlr-kotlin matching issue.
Current parser passes `operationUses operationUse` as *left* candidate of
`operationUses` branch (which is `operationUse | operationUses operationUse`).
It is likely due to https://github.com/Strumenta/antlr-kotlin/issues/22 and
there is no fix yet.

A workaround is added to the visitor implementation that handles those
bogus inputs to the method gracefully (should be only a workaround until
a real fix lands on antlr-kotlin...).

This makes existing tests pass.