IR improvements, but Go back to LEGACY.
... as IR fails to compile some existing MMLs.

A lot of IR problems are resolved in this changeset. Missing JsExports are
filled. Companion objects are not generated so they are now invoked via
some non-static functions or global functions (they turned out to work fine).
JsInteropSupport class is gone (we moved back to global functions).

On vscode-extension side, our module loader code is more complicated to
handle both LEGACY and IR (for future experiment). We avoid Companion
objects so that it wouldn't block us when we switch to IR.