migrate to ktmidi 0.6.0 API. Adjust some JsExports. Minor fix in compiler tool.
The API change is huge, if we want to catch up with it and eliminate the
related warnings.

JsExport was part of the warnings. It's still there but less.

The compiler console tool had both --midi2 and --midi2x, but they differed
only in deltaTime handling, which is outdated with June 2023 updates anymore.